gpt-recruit-front/pyproject.toml
[tool.poetry]
name = "gpt-recruit-front"
version = "0.1.0"
description = "ai 자소서 프론트"
authors = ["Ruo Lee <comsa33@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.34.0"
pymongo = "^4.7.2"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
icecream = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
